+++ limen/__main__.py
@@ -1,0 +1,5 @@
+from limen.cli.main import cli
+
+
+if __name__ == '__main__':
+    cli(prog_name='python -m limen')